backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-trip-planner-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "Django>=5.0,<5.1",
  "djangorestframework>=3.15,<3.16",
  "djangorestframework-simplejwt>=5.3,<5.4",
  "drf-spectacular>=0.27,<0.28",
  "django-environ>=0.11,<0.12",
  "django-cors-headers>=4.3,<4.4",
  "psycopg2-binary>=2.9,<3.0",
  "gunicorn>=21.2,<22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "pytest-django>=4.8,<5.0",
  "model-bakery>=1.17,<2.0",
  "black>=24.0,<25.0",
  "ruff>=0.5,<0.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["config", "config.*", "apps", "apps.*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "B", "UP"]
